I'm doing a lot of reductions starting with full-sized IR modules, trying to find minimal triggers for random issues detected by alive2. Even after llvm-reduce reaches a fixpoint, there's often some control-flow-related junk, often towards the end of the module, that isn't getting cleaned up by existing passes that we have.
This patch adds three reduction passes. The first one tries to make
every conditional branch into an unconditional branch to the original
branch's true target. The second pass does the same thing, but uses the
false target. The third pass targets unconditional branches. For a BB
terminated by an unconditional branch, all instructions in the BB are
sunk to the start of the successor block and all branches targeting
the original block are pointed to the successor. Then we delete the
block.

On a collection of 53 IR files that I'm using to benchmark
llvm-reduce, this patch changes the average final file size from 710
bytes to 615 bytes, about a 15% improvement.